Fixed simulation, relative ankle angle
changed relative ankle angle.
Found some drift in the "keep arm down" condition of the control in hybrid_sim... when g and k = 0, it cuased led to still drift some.
Updated ankle angle limit.
checked it worked with k and g set to zero and nonzero values (in all permutations)