Fixed derive, first "successful" param sweep
Potential energy term was messed up after changing theta-ankle to relative. Fixed.
Param sweep still noisy, I think because of (a) the system being chaotic naturally, with lost of different ways to "resonate" and (b) the arm swing is limited by the max motor torque (would be good to check that).